# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Wallet Service
  description: Service for APIs for wallet transactions
  version: ""
host: localhost:8088
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /admin/wallets:
    get:
      tags:
      - v1/wallet
      summary: admin/wallets v1/wallet
      description: This API returns all wallets
      operationId: v1/wallet#admin/wallets
      parameters:
      - name: Authorization
        in: header
        description: JWT token used to perform authorization
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/V1WalletAdminWalletsResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /balance:
    get:
      tags:
      - v1/wallet
      summary: balance v1/wallet
      description: This API checks balance in wallet
      operationId: v1/wallet#balance
      parameters:
      - name: userID
        in: query
        description: user ID
        required: true
        type: integer
        minimum: 0
      - name: Authorization
        in: header
        description: JWT token used to perform authorization
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/V1WalletBalanceResponseBody'
            required:
            - Amount
            - UpdatedAt
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /deposit:
    post:
      tags:
      - v1/wallet
      summary: deposit v1/wallet
      description: This API deposit money in wallet
      operationId: v1/wallet#deposit
      parameters:
      - name: Authorization
        in: header
        description: JWT token used to perform authorization
        required: false
        type: string
      - name: DepositRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/V1WalletDepositRequestBody'
          required:
          - amount
      responses:
        "201":
          description: Created response.
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /health:
    get:
      tags:
      - v1/wallet
      summary: health v1/wallet
      description: This API checks for status 200 for downstream services
      operationId: v1/wallet#health
      responses:
        "200":
          description: OK response.
      schemes:
      - http
  /transfer:
    post:
      tags:
      - v1/wallet
      summary: transfer v1/wallet
      description: This API transfer money from one wallet to another
      operationId: v1/wallet#transfer
      parameters:
      - name: Authorization
        in: header
        description: JWT token used to perform authorization
        required: false
        type: string
      - name: TransferRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/V1WalletTransferRequestBody'
          required:
          - receiverID
          - amount
      responses:
        "201":
          description: Created response.
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /withdraw:
    post:
      tags:
      - v1/wallet
      summary: withdraw v1/wallet
      description: This API withdraw money from wallet
      operationId: v1/wallet#withdraw
      parameters:
      - name: Authorization
        in: header
        description: JWT token used to perform authorization
        required: false
        type: string
      - name: WithdrawRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/V1WalletWithdrawRequestBody'
          required:
          - amount
      responses:
        "201":
          description: Created response.
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
definitions:
  V1WalletAdminWalletsResponseBody:
    title: V1WalletAdminWalletsResponseBody
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/WalletResponseBody'
        example:
        - Amount: 0.6964249561747816
          UpdatedAt: Sed voluptate repellendus.
          userID: 7307358692524276138
        - Amount: 0.6964249561747816
          UpdatedAt: Sed voluptate repellendus.
          userID: 7307358692524276138
    example:
      data:
      - Amount: 0.6964249561747816
        UpdatedAt: Sed voluptate repellendus.
        userID: 7307358692524276138
      - Amount: 0.6964249561747816
        UpdatedAt: Sed voluptate repellendus.
        userID: 7307358692524276138
      - Amount: 0.6964249561747816
        UpdatedAt: Sed voluptate repellendus.
        userID: 7307358692524276138
      - Amount: 0.6964249561747816
        UpdatedAt: Sed voluptate repellendus.
        userID: 7307358692524276138
  V1WalletBalanceResponseBody:
    title: V1WalletBalanceResponseBody
    type: object
    properties:
      Amount:
        type: number
        example: 0.7818719061927136
        format: double
      UpdatedAt:
        type: string
        example: Non numquam provident sunt excepturi.
    example:
      Amount: 0.912886180732953
      UpdatedAt: Ratione et accusantium mollitia.
    required:
    - Amount
    - UpdatedAt
  V1WalletDepositRequestBody:
    title: V1WalletDepositRequestBody
    type: object
    properties:
      amount:
        type: number
        example: 0.2397280600809405
        minimum: 0
    example:
      amount: 0.49603931963317643
    required:
    - amount
  V1WalletTransferRequestBody:
    title: V1WalletTransferRequestBody
    type: object
    properties:
      amount:
        type: number
        example: 0.5840829876924009
        minimum: 0
      receiverID:
        type: integer
        example: 6428160049097336618
        minimum: 0
    example:
      amount: 0.7261107977239808
      receiverID: 9022785266939615928
    required:
    - receiverID
    - amount
  V1WalletWithdrawRequestBody:
    title: V1WalletWithdrawRequestBody
    type: object
    properties:
      amount:
        type: number
        example: 0.18242176752907413
        minimum: 0
    example:
      amount: 0.9924694802550078
    required:
    - amount
  WalletResponseBody:
    title: WalletResponseBody
    type: object
    properties:
      Amount:
        type: number
        example: 0.38049781781858366
        format: double
      UpdatedAt:
        type: string
        example: Et quidem qui quia nobis rerum necessitatibus.
      userID:
        type: integer
        example: 1961174429351993395
        format: int64
    example:
      Amount: 0.9106736198455511
      UpdatedAt: Omnis ea et delectus.
      userID: 5460122318273755654
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    name: Authorization
    in: header